Some refactoring of the CSV example and also improved the README file for the CSV example
